ExoData object and constructors

ExoData is a subclass of GenomicRanges, used to asses the quality of ChIP-exo/nexus sample.

ExoData(file = NULL, reads = NULL, height = 1,
  mc.cores = getOption("mc.cores", 2L), save.reads = FALSE,
  nregions = 1000, ntimes = 100, verbose = TRUE)

Arguments

file

a character value with location of the bam file with the aligned reads.

reads

a GAlignments object with the aligned reads of a ChIP-exo sample. It is meant to be used instead of file.

height

a numeric value indicating the value used to slice the coverage of the experiment into a set of regions.

mc.cores

a numeric value with the number of cores to use, i.e. at most how many child processes will be run simultaneously.

save.reads

a logical value to indicate if the reads are stored in the ExoData object. The default value is FALSE.

nregions

a numeric value indicating the number of regions sampled to estimate the quality parameter distributions. The default value is 1e3.

ntimes

a numeric value indicating the number of times that regions are sampled to estimate the quality parameter distributions. The default value is 1e2.

verbose

a logical value indicating if the user want to receive progress details. The default value is FALSE.

Value

It returns an ExoData object with the regions obtained after partitioning the genome and the summary statistics for each region. If the save.reads parameter is TRUE then it contains a GRanges object with the reads of the ChIP-exo experiment.

Examples

files <- list.files(system.file("extdata",package = "ChIPexoQualExample"),
    full.names = TRUE)
ExoData(files[5],mc.cores = 2L)

#> Using 'file' argument

#> Creating ExoData object using aligned reads in /home/rene/R/x86_64-pc-linux-gnu-library/3.3/ChIPexoQualExample/extdata/ChIPexo_carroll_FoxA1_mouse_rep3_chr1.bam

#> Keeping reads in object: No

#> Loading experiment reads

#> Calculating summary statistics

#> Calculating quality scores distribution

#> Done!

#> ExoData object with 215936 ranges and 11 metadata columns:
#>            seqnames                 ranges strand |  fwdReads  revReads
#>               <Rle>              <IRanges>  <Rle> | <integer> <integer>
#>        [1]     chr1     [3001926, 3001961]      * |         6         0
#>        [2]     chr1     [3002412, 3002447]      * |         0         7
#>        [3]     chr1     [3003987, 3004022]      * |         0        11
#>        [4]     chr1     [3006189, 3006224]      * |         6         0
#>        [5]     chr1     [3012445, 3012480]      * |         0         1
#>        ...      ...                    ...    ... .       ...       ...
#>   [215932]     chr1 [197189055, 197189090]      * |        10         0
#>   [215933]     chr1 [197189266, 197189301]      * |         2         0
#>   [215934]     chr1 [197189679, 197189714]      * |         9         0
#>   [215935]     chr1 [197192288, 197192323]      * |         0         4
#>   [215936]     chr1 [197192797, 197192832]      * |         3         0
#>               fwdPos    revPos     depth uniquePos                ARC
#>            <integer> <integer> <integer> <integer>          <numeric>
#>        [1]         1         0         6         1  0.166666666666667
#>        [2]         0         1         7         1  0.194444444444444
#>        [3]         0         1        11         1  0.305555555555556
#>        [4]         1         0         6         1  0.166666666666667
#>        [5]         0         1         1         1 0.0277777777777778
#>        ...       ...       ...       ...       ...                ...
#>   [215932]         1         0        10         1  0.277777777777778
#>   [215933]         1         0         2         1 0.0555555555555556
#>   [215934]         1         0         9         1               0.25
#>   [215935]         0         1         4         1  0.111111111111111
#>   [215936]         1         0         3         1 0.0833333333333333
#>                           URC       FSR         M         A
#>                     <numeric> <numeric> <numeric> <numeric>
#>        [1]  0.166666666666667         1      -Inf       Inf
#>        [2]  0.142857142857143         0      -Inf      -Inf
#>        [3] 0.0909090909090909         0      -Inf      -Inf
#>        [4]  0.166666666666667         1      -Inf       Inf
#>        [5]                  1         0      -Inf      -Inf
#>        ...                ...       ...       ...       ...
#>   [215932]                0.1         1      -Inf       Inf
#>   [215933]                0.5         1      -Inf       Inf
#>   [215934]  0.111111111111111         1      -Inf       Inf
#>   [215935]               0.25         0      -Inf      -Inf
#>   [215936]  0.333333333333333         1      -Inf       Inf
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths